Adds server.rs, rocket_routes. Starts declaring api routes, auth info, repositories

+++ src/models.rs
@@ -2,9 +2,11 @@ use std::{io::Write, str::FromStr};
 
 use bigdecimal::BigDecimal;
 use chrono::NaiveDateTime;
-use diesel::{deserialize::{FromSql, FromSqlRow}, expression::AsExpression, pg::{Pg, PgValue}, prelude::{Associations, Identifiable}, serialize::{Output, ToSql}, sql_types::{Double, Text}, AsChangeset, Insertable, Queryable};
+use diesel::{deserialize::{FromSql, FromSqlRow}, expression::AsExpression, pg::{Pg, PgValue}, prelude::{Associations, Identifiable}, serialize::{Output, ToSql}, sql_types::{Double, Numeric, Text}, AsChangeset, Insertable, Queryable};
 use rocket::serde::{Deserialize, Serialize};
 use crate::schema::*;
+
+use self::items::columns;
 
 #[derive(Serialize, Deserialize, Queryable, Identifiable, AsChangeset)]
 pub struct Item {
@@ -12,7 +14,7 @@ pub struct Item {
     pub id: i32,
     pub name: String,
     pub description: Option<String>,
-    pub price: BigDecimal,
+    pub price: f32,
     #[serde(skip_deserializing)]
     pub created_at: Option<NaiveDateTime>,
     pub quantity: i32,
@@ -23,6 +25,6 @@ pub struct Item {
 pub struct NewItem {
     pub name: String,
     pub description: Option<String>,
-    pub price: BigDecimal,
+    pub price: Numeric,
     pub quantity: i32,
 }
